pyproject.toml
[tool.poetry]
name = "arcaflow-plugin-stressng"
version = "0.1.0"
description = "arcaflow plugin for stressng"
authors = ["arcalot"]
license = "Apache-2.0+GPL-2.0-only"
readme = "README.md"

packages = [
   { include="stressng_plugin.py", from="./arcaflow_plugin_stressng"  },
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.4"
scipy = "^1.9.3"
pyyaml = ">=6,<6.1"
arcaflow-plugin-sdk = "^0.14.0b1"

[tool.poetry.dev-dependencies]
docformatter = "^1.5.0"
autoflake = "^2.0.0"
pydocstyle = "^6.1.1"
black = "^24.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
